All cart to be restored without deleting stored copy
Helps with persistent cart storage

+++ src/Cart.php
@@ -369,7 +369,7 @@ class Cart
      * @param mixed $identifier
      * @return void
      */
-    public function restore($identifier)
+    public function restore($identifier, $delete_store = true)
     {
         if( ! $this->storedCartWithIdentifierExists($identifier)) {
             return;
@@ -396,8 +396,10 @@ class Cart
 
         $this->instance($currentInstance);
 
-        $this->getConnection()->table($this->getTableName())
-            ->where('identifier', $identifier)->delete();
+        if ($delete_store === true) {
+            $this->getConnection()->table($this->getTableName())
+                ->where('identifier', $identifier)->delete();
+        }
     }
 
     /**
